+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.0.3 (Released 2024-10-14)
+    * Add decoder for multiple optional fields with dictionaries
+
+### 0.0.2 (Released 2024-7-11)
+    * Add overflow decoding logic
+
 ### 0.0.1-alpha (Released 2024-6-20)
     * First open alpha release of YAMLicious. Rough feature set:
     * Tokenize YAML files